write_to_device: improvements on read & write img
enforced 2 requirements:
* Only full buffers are read and written to a device
* The exact amount of image bytes are written. This implies that the
  buffer size of the last write will be trimmed.

Note: there is an implicit requiment hidden in this code. The buffersize
that is written needs to be a multiple of the sector size of the device
that is written to.

+++ tpi_rs/src/middleware/firmware_update/mod.rs
@@ -121,7 +121,7 @@ pub enum FlashStatus {
     Idle,
     Setup,
     Progress {
-        read_percent: u64,
+        read_percent: usize,
         est_minutes: u64,
         est_seconds: u64,
     },
